+++ compliance_controller/wrench_to_joint_vel_pub/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(wrench_to_twist_pub)
+project(wrench_to_joint_vel_pub)
 
 add_compile_options(-std=c++11)
 
@@ -44,9 +44,9 @@ add_library(compliant_control
   src/${PROJECT_NAME}/compliant_control.cpp
 )
 
-add_executable(wrench_to_twist_pub src/${PROJECT_NAME}/wrench_to_twist_pub.cpp)
-add_dependencies(wrench_to_twist_pub ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
-target_link_libraries(wrench_to_twist_pub ${catkin_LIBRARIES} compliant_control ${Eigen_LIBRARIES})
+add_executable(wrench_to_joint_vel_pub src/${PROJECT_NAME}/wrench_to_joint_vel_pub.cpp)
+add_dependencies(wrench_to_joint_vel_pub ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+target_link_libraries(wrench_to_joint_vel_pub ${catkin_LIBRARIES} compliant_control ${Eigen_LIBRARIES})
 
 add_executable(test_compliance_publisher src/${PROJECT_NAME}/test_compliance_publisher.cpp)
 add_dependencies(test_compliance_publisher ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
@@ -61,7 +61,7 @@ install(DIRECTORY launch/
   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/launch
 )
 
-install(TARGETS compliant_control wrench_to_twist_pub
+install(TARGETS compliant_control wrench_to_joint_vel_pub
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
